Split function declarations from functional code
